refactor(dev-menu): relocate to src/private/utilities
