+++ packages/react-native/index.js
@@ -85,7 +85,7 @@ import typeof UTFSequence from './Libraries/UTFSequence';
 import typeof * as Appearance from './Libraries/Utilities/Appearance';
 import typeof BackHandler from './Libraries/Utilities/BackHandler';
 import typeof DeviceInfo from './Libraries/Utilities/DeviceInfo';
-import typeof DevMenu from './Libraries/Utilities/DevMenu';
+import typeof DevMenu from './src/private/utilities/DevMenu';
 import typeof DevSettings from './Libraries/Utilities/DevSettings';
 import typeof Dimensions from './Libraries/Utilities/Dimensions';
 import typeof PixelRatio from './Libraries/Utilities/PixelRatio';
@@ -240,7 +240,7 @@ module.exports = {
     return require('./Libraries/Utilities/DeviceInfo');
   },
   get DevMenu(): DevMenu {
-    return require('./Libraries/Utilities/DevMenu');
+    return require('./src/private/utilities/DevMenu');
   },
   get DevSettings(): DevSettings {
     return require('./Libraries/Utilities/DevSettings');
